TIP-1167: add tests on the query to get connector products
