+++ src/Akeneo/Pim/Enrichment/Bundle/Product/Query/Sql/GetProductAssociationsByProductIdentifiers.php
@@ -103,7 +103,9 @@ SQL;
 
             $filteredAssociations = [];
             foreach ($associations as $associationType => $productAssociations) {
-                $filteredAssociations[$associationType]['products'] = array_filter($productAssociations);
+                $association = array_values(array_filter($productAssociations));
+                sort($association);
+                $filteredAssociations[$associationType]['products'] = $association;
             }
 
             $results[$row['product_identifier']] = $filteredAssociations;
