+++ tests/back/Pim/Enrichment/EndToEnd/Product/Product/ExternalApi/ListProducts/ErrorListProductEndToEnd.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace AkeneoTest\Pim\Enrichment\EndToEnd\Product\Product\ExternalApi;
+namespace AkeneoTest\Pim\Enrichment\EndToEnd\Product\Product\ExternalApi\ListProducts;
 
 use Akeneo\Test\Integration\Configuration;
 use Symfony\Bundle\FrameworkBundle\Client;
